# Conflicts:
#	src/main/java/pengyi/application/report/command/CreateReportCommand.java
#	src/main/java/pengyi/domain/service/report/ReportService.java

+++ src/main/java/pengyi/domain/service/user/driver/DriverService.java
@@ -283,7 +283,9 @@ public class DriverService implements IDriverService {
 
         driverRepository.update(driver);
         fileUploadService.move(headPic.substring(headPic.lastIndexOf("/") + 1));
-        fileUploadService.delete(oldHeadPic.substring(oldHeadPic.lastIndexOf("/") + 1));
+        if (null != oldHeadPic) {
+            fileUploadService.delete(oldHeadPic.substring(oldHeadPic.lastIndexOf("/") + 1));
+        }
         return driver;
     }
 }
